🚧 WIP: Gestion des messages
