+++ assets/app.js
@@ -21,7 +21,7 @@ import './js/zuck-stories.js'
 import './vendor/bootstrap/dist/js/bootstrap.min.js'
 import './vendor/flatpickr/dist/flatpickr.js'
 import './vendor/glightbox-master/dist/js/glightbox.min.js'
-import './vendor/overlayscrollbars/overlayscrollbars.index.js'
+import './vendor/OverlayScrollbars-master/js/OverlayScrollbars.min.js'
 import './vendor/choices/choices.index.js'
 import './vendor/colors/colors.index.js'
 import './vendor/zuck/dist/zuck.css.min.js'
