+++ templates/base.html.twig
@@ -107,6 +107,7 @@
         <script src="{{ asset('vendor/tiny-slider/tiny-slider.js') }}"></script>
         <script src="{{ asset('vendor/glightbox-master/dist/js/glightbox.min.js') }}"></script>
         <script src="{{ asset('vendor/plyr/plyr.js') }}"></script>
+        <script src="{{ asset('vendor/OverlayScrollbars-master/js/OverlayScrollbars.min.js') }}"></script>
 
         <!-- Theme Functions -->
         <script src="{{ asset('js/functions.js') }}"></script>
