+++ templates/home/index.html.twig
@@ -78,7 +78,7 @@
                                                                     <i class="fas fa-age"></i>
                                                                     {{ light.age }}
                                                                 </span>
-                                                                <p class="m-0">ans</p>
+                                                                <p class="h6 text-end fw-lighter">ans</p>
                                                             </div>
                                                         </div>
 
@@ -112,7 +112,7 @@
                                                         </div>
                                                         <div class="d-flex pt-1">
                                                             {# TODO: Ecrire un message #}
-                                                            <a href="#" class="btn btn-outline-primary me-2 flex-grow-1">
+                                                            <a href="{{ path('app_messaging_index') }}" class="btn btn-outline-primary me-2 flex-grow-1">
                                                                 <i class="fas fa-envelope"></i>
                                                                 Lui écrire
                                                             </a>
