+++ templates/messaging/_light_photo.html.twig
@@ -1,0 +1,13 @@
+{% if light.photo is not same as 'default.jpg' %}
+    <img src="{{ asset('uploads/photos/' ~ light.photo) }}"
+         alt="miniature {{ light.fullname }}"
+         class="avatar-img rounded-circle"
+         style="max-height: 180px; max-width: 180px; border-radius: 10px;"
+    >
+{% else %}
+    <img src="{{ asset('uploads/photos/default/' ~ light.photo) }}"
+         alt="miniature par défaut"
+         class="avatar-img rounded-circle"
+         style="max-height: 180px; max-width: 180px; border-radius: 10px;"
+    >
+{% endif %}
